safari-534.53-branch/LayoutTests/svg/animations/animate-inherit-css-property-expected.txt

CONSOLE MESSAGE: line 2: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

Tests animation with 'inherit'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
